Fix flaky tests in DruidCoordinatorTest

+++ server/src/test/java/org/apache/druid/server/coordinator/DruidCoordinatorTest.java
@@ -674,17 +674,16 @@ public class DruidCoordinatorTest extends CuratorTestBase
     pathChildrenCache.getListenable().addListener(
         (CuratorFramework client, PathChildrenCacheEvent event) -> {
           if (CuratorUtils.isChildAdded(event)) {
-              DataSegment segment = findSegmentRelatedToCuratorEvent(segments, event);
-              if (segment != null) {
-                if (countDownLatch.getCount() > 0) {
-                  server.addDataSegment(segment);
-                  curator.delete().guaranteed().forPath(event.getData().getPath());
-                  countDownLatch.countDown();
-                } else {
-                  Assert.fail("The segment path " + event.getData().getPath()
-                              + " is assigned to the server more times than expected. Expected only " + latchCount);
-                }
+            DataSegment segment = findSegmentRelatedToCuratorEvent(segments, event);
+            if (segment != null && server.getSegment(segment.getId()) == null) {
+              if (countDownLatch.getCount() > 0) {
+                server.addDataSegment(segment);
+                curator.delete().guaranteed().forPath(event.getData().getPath());
+                countDownLatch.countDown();
+              } else {
+                Assert.fail("The segment path " + event.getData().getPath() + " is not expected");
               }
+            }
           }
         }
     );
